src/api/requirements.txt
# Base packages
cachetools==6.0.0
python-dotenv==1.1.0
fastapi==0.115.12
uvicorn[standard]
pydantic[email]

# Azure SDK Core
azure-core==1.34.0
requests==2.32.3
types-requests==2.32.4.20250611
aiohttp==3.12.9

# Azure Services
azure-identity==1.23.0
azure-search-documents==11.6.0b12
azure-ai-projects==1.0.0b12
azure-ai-inference==1.0.0b9
azure-cosmos==4.9.0

# Additional utilities
semantic-kernel[azure]==1.32.2
openai==1.84.0
pyodbc==5.2.0
pandas==2.3.0

opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc
azure-monitor-events-extension
opentelemetry-sdk==1.31.1
opentelemetry-api==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-instrumentation==0.52b1
azure-monitor-opentelemetry==1.6.8

# Development tools
pytest==8.3.5
pytest-cov==6.1.1
pytest-asyncio==0.26.0
